feat(Input): always display the error tooltip when the input is focused
This commit changes the tabIndex of the tooltip Container from 0 to -1, the element is focusable by the keyboard and the cursor when the tabIndex is set to 0 and only by the keyboard when set to -1.
There is no need to make the Container element focusable by the keyboard now that the tooltip will be displayed automatically when the input is focused.